Update docs README with build instructions using tox
None of the existing commands built the docs properly for me (I had to
revert the changes in 901892d to docs/conf.py to get them to build
properly, and even then no tooltips displayed).

It seems that scrapy is now using tox to build the docs, so I have
replaced the commands with the tox one that worked for me.